Support showing trips without starting time or location
Current solution is to pretend that end time/station is start time/station
if no real start data is available. It's confusing as there is no
distinguishing between start and end visually and hence it can be assumed
it's a start of a journey rather than the end. It's tolerable for occasional
hiccup but Shenzhen Tong has only ending stations AFAICT.